- Fixed some warnings on MSVC
git-svn-id: http://svn.code.sf.net/p/irrrpgbuilder/code@52 528fbfa9-60f6-4f4a-a82b-d133dc65ca33

+++ src/EditorCamera.cpp
@@ -76,8 +76,8 @@ int EditorCamera::getCamera()
 
 void EditorCamera::setCameraHeight(irr::f32 increments)
 {
-	int max = 0;
-	int min = 0;
+	f32 max = 0;
+	f32 min = 0;
 	cameraHeight-=(increments/10);
 	switch (camera)
 	{
